% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj2graph.R
\name{adj2graph}
\alias{adj2graph}
\title{Visualize a graph from its adjacency matrix}
\usage{
adj2graph(adj, ...)
}
\arguments{
\item{adj}{An adjacency matrix (a matrix with rows and columns representing nodes and arguments representing an edge or edge weight between two specific nodes).}

\item{...}{Any additional arguments.}
}
\value{
A list including the desired features of a graph.
}
\description{
Visualizes a graph from its adjacency matrix.
}
\section{Additional arguments}{

\describe{
\item{directed}{A logical indicating if the graph is directed or not.}
\item{community}{A logical indicating if the node communities should be found and the nodes should be colored by them.}
\item{plot}{A logical showing if the graph should be visualized statically (with qgraph).}
\item{vis_net}{A logical showing if the graph should be visualized interactively (with visNetwork).}
\item{betweenness}{While visualizing interactively a logical showing if the node sizes should be proportional to their betweenness centrality.}
\item{log}{A logical indicating whether to log the computation times.}
\item{verbose}{A logical indicating whether to print out the computation steps.}
}
}

\author{
Elyas Heidari
}
