% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{china_rt_estimates}
\alias{china_rt_estimates}
\title{Daily estimates of effective reporduction number from COVID-19 case data in China}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 497 rows and 5 columns.}
\usage{
china_rt_estimates
}
\description{
A dataset of effective reproduction number (Rt) estimates over time in different provinces of China.
}
\keyword{datasets}
