% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexp_net.R
\name{coexp_net}
\alias{coexp_net}
\title{Coexpression network based on the mutual information}
\usage{
coexp_net(mtx, ...)
}
\arguments{
\item{mtx}{An expression or count matrix with rows as features (e.g., genes) and columns as observations (e.g., cells).}

\item{...}{Any additional arguments.}
}
\value{
The adjacency matrix as a sparse matrix of type \code{dgCMatrix}.
}
\description{
Computes mutual information of pairs of features (e.g., genes), to construct an interaction network based on high mutual informations.
}
\details{
Running time might be long and using multiple cores is recommended.
}
\section{Additional arguments}{

\describe{
\item{mc.cores}{Number of cores to be used for parallelization.}
\item{num_edges}{A numeric indicating the final number of edges (top n edges to be picked).}
\item{min_mi}{Minimum of the mutual information tolerable (if \code{num_edges = NULL}).}
\item{min_rel_prop}{Minimum of the mutual information tolerable relative to the maximum among all edges(if \code{num_edges = NULL}).}
\item{log}{A logical indicating whether to log the computation times.}
\item{verbose}{A logical indicating whether to print out the computation steps.}
}
}

\author{
Elyas Heidari
}
