% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fisherLogseries.R
\name{dfish}
\alias{dfish}
\alias{pfish}
\title{Fisher's logseries}
\usage{
dfish(x, beta, log = FALSE)

pfish(q, beta, log = FALSE)
}
\arguments{
\item{x}{}

\item{beta}{}

\item{log}{}

\item{n}{}
}
\value{
A numeric vector of length equal to the input
}
\description{
\code{dfish} gives the probability mass function, \code{pfish} gives the cumulative mass
function, \code{qfish} the quantile function, \code{rfish} randome number generation
}
\details{
These functions assume infinite support of the Fisher logseries
}
\examples{
dfish(1:10, 0.01)
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{Fisher}
\keyword{SAD}
\keyword{abundance,}
\keyword{logseries,}
\keyword{species}

