% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exante_movement_data}
\alias{exante_movement_data}
\title{Example exante movement data}
\format{A data frame}
\usage{
exante_movement_data
}
\description{
Example exante movement data
}
\keyword{datasets}
