% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitSAD.R
\name{fitSAD}
\alias{fitSAD}
\title{Fit a model for a species abundance distribution}
\usage{
fitSAD(x, models = c("fish", "plnorm", "stick", "tnegb", "tpois"))
}
\arguments{
\item{x}{vector of integers representing a sample of species abundances}

\item{models}{character vector of models to be fit (can be more than one model, see examples)}
}
\value{
An object of class \code{logLik} with attributes "\code{df}" and "\code{nobs}."
}
\description{
\code{fitSAD} uses maximum likelihood to fit one or several SAD models to data
}
\details{
The default behavior fits all models availible in \code{pika}
}
\examples{
dplnorm(1:10, 0.5, 0.1)
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\seealso{
logLik
}
\keyword{Maximum}
\keyword{SAD}
\keyword{abundance,}
\keyword{likelihood,}
\keyword{species}

