% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistFuns.R
\name{getdfun}
\alias{getdfun}
\alias{getpfun}
\alias{getqfun}
\alias{getrfun}
\title{Functions to get distribution functions from objects of class \code{sad}}
\usage{
getdfun(s)

getpfun(s)

getqfun(s)

getrfun(s)
}
\arguments{
\item{s}{an object of class \code{sad}}

\item{...}{arguments to be passed to distribution functions, e.g. \code{log}, \code{lower.tail}}
}
\value{
A function (d, p, q, r) parameterized with the distribution of the \code{sad} object
}
\description{
Get distribution functions from objects of class \code{sad}
}
\details{
See examples
}
\examples{
x <- rtpois(1000, 2)
sad1 <- fitSAD(x, models='tpois')
getdfun(sad1[[1]])(1:10, log=TRUE)

}
\seealso{
fitSAD
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{SAD}
\keyword{abundance,}
\keyword{species}
