% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_net.R
\name{ggm_net}
\alias{ggm_net}
\title{Gaussian Graphical Model}
\usage{
ggm_net(mtx, ...)
}
\arguments{
\item{mtx}{An expression or count matrix with rows as features (e.g., genes) and columns as observations (e.g., cells).}

\item{...}{Any additional arguments.}
}
\value{
The adjacency matrix as a sparse matrix of type \code{dgCMatrix}.
}
\description{
Fits a Gaussian Graphical Model to continuous-valued dataset employing a subset of methods from stepwise AIC, stepwise BIC, stepwise significance test, partial correlation thresholding, edgewise significance test, or glasso.
}
\details{
The function combines the methods to construct the model, that is, the edge set is the intersection of all edge sets each of which is found by a method. The package gRim is used to implement AIC, BIC, and stepwise significance test. The method glasso from the package glasso is used to provide a sparse estimation of the inverse covariance matrix.
}
\section{Additional arguments}{

\describe{
\item{method}{A string or list of strings indicate methods used to construct the model. See the details for more information. (default = 'glasso')}
\item{stability}{A string or list of strings indicate methods used to construct the model. See the details for more information. (default = 'glasso')}
\item{rho_vals}{A vector of non-negative numerics, including regularization parameter for glasso (default = 0.1). To be used only when the method 'glasso' is used.}
\item{mc.cores}{Number of cores to be used for parallelization.}
\item{threshold}{A threshold for partial correlation thresholding method (default = 0.05). To be used only when the method 'threshold' is used.}
\item{significance}{A cutoff for edge significance (default = 0.05). To be used only when the method 'significance' is used.}
\item{log}{A logical indicating whether to log the computation times.}
\item{verbose}{A logical indicating whether to print out the computation steps.}
}
}

\references{
Højsgaard, S., Edwards, D., & Lauritzen, S. (2012). Graphical Models with R. Springer US. \url{https://doi.org/10.1007/978-1-4614-2299-0}

Friedman, J., Hastie, T., & Tibshirani, R. (2007). Sparse inverse covariance estimation with the graphical lasso. Biostatistics, 9(3), 432–441. \url{https://doi.org/10.1093/biostatistics/kxm045}

Abreu, G. C. G., Edwards, D., & Labouriau, R. (2010). High-Dimensional Graphical Model Search with thegRapHDRPackage. Journal of Statistical Software, 37(1). \url{https://doi.org/10.18637/jss.v037.i01}
}
\author{
Elyas Heidari
}
