% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_col_act.R
\name{graph_col_act}
\alias{graph_col_act}
\title{A wrapper for \code{igraph::plot.igraph} for node activities (continuous color values)}
\usage{
graph_col_act(graph, values, contrast = 1, lay, title)
}
\arguments{
\item{graph}{An igraph object.}

\item{values}{A numeric vector of the same length as the number of nodes indicating node values.}

\item{contrast}{A positive numeric indicating the contrast of node colors (the higher the value the more the contrast). default=1.}

\item{lay}{An igraph valid network layout. default = 'NULL'.}

\item{title}{A string indicating the title of the graph plot.}
}
\value{
An igraph plot.
}
\description{
Visualizes a graph and colors its nodes based on their values (e.g., activities).
}
\author{
Elyas Heidari
}
