% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_experiment.R
\name{load_experiment}
\alias{load_experiment}
\title{Load a single cell experiment}
\usage{
load_experiment(base_dir, tag, ...)
}
\arguments{
\item{base_dir}{A string indicating the experiment's base directory.}

\item{tag}{A string indicating the experiment's tag. Should be name of a folder containing train and test folders each of which including \code{counts.txt}, \code{cells.txt}, \code{genes.txt}.}

\item{...}{Any additional arguments.}
}
\value{
A SingleCellExperiment with:
\describe{
\item{rowData(.)$is_TF}{indicating if the gene corresponding to the row is a transcription factor.}
\item{rowData(.)$gene_var}{Modeled variance of the log-expression profiles for each gene, based on \code{scran::modelGeneVar}.}
}
}
\description{
Loads an experiment and write it into a SingleCellExperiment object. Also preprocesses data with \code{scran}
}
\details{
\code{counts.txt} contains the counts matrix, \code{cells.txt} contains cell ids, \code{genes.txt} contains gene ids.
}
\section{Additional arguments}{

\describe{
\item{spec}{A string indicating the sapience. possible values = \code{c('cattle', 'chicken','chimpanzee', 'dog_domestic', 'frog_western clawed', 'human', 'mouse_laboratory','rat', 'zebrafish')}.}
\item{log}{A logical indicating whether to log the computation times.}
\item{verbose}{A logical indicating whether to print out the computation steps.}
}
}

\author{
Elyas Heidari
}
