% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sadMethods.R
\name{logLik.sad}
\alias{logLik.sad}
\title{Log likelihood method for objects of class \code{sad}}
\usage{
\method{logLik}{sad}(x)
}
\arguments{
\item{x}{an object of class \code{sad}}
}
\value{
An object of class \code{logLik} with attributes "\code{df}" and "\code{nobs}".
}
\description{
Compute log likelihood for objects of class \code{sad}
}
\details{
See examples
}
\examples{
x <- rtpois(1000, 2)
sad1 <- fitSAD(x, models='tpois')
logLik(sad1[[1]])
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\seealso{
fitSAD
}
\keyword{SAD}
\keyword{abundance,}
\keyword{species}

