% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logseries.R
\name{dlseries}
\alias{dlseries}
\alias{plseries}
\alias{qlseries}
\alias{rlseries}
\title{Logseries distribution}
\usage{
dlseries(x, beta, log = FALSE)

plseries(q, beta, lower.tail = TRUE, log = FALSE)

qlseries(p, beta, lower.tail = TRUE, log = FALSE)

rlseries(n, beta)
}
\arguments{
\item{x}{vector of integers for which to return the probability}

\item{beta}{parameter of the log series, see Details}

\item{log}{logical, should the log probability be used}

\item{q}{vector of integers for which to return the cumulative probability}

\item{lower.tail}{logical, should the lower tail be used}

\item{p}{vector of probabilities for which to return the quantile}

\item{n}{number of random replicates}
}
\value{
A numeric vector of length equal to the input
}
\description{
\code{dlseries} gives the probability mass function, \code{plseries} gives the cumulative mass 
function, \code{qlseries} the quantile function, \code{rlseries} randome number generation
}
\details{
These functions assume infinite support of the logseries from [1, Inf).  An alternate parameterization is used to allow a larger range (i.e. the positive number line) of the parameter \code{beta}: \eqn{\frac{1}{log(1/(1-e^{-\beta}))} * \frac{e^{-\beta x}}{x}}.
}
\examples{

dlseries(1:10, 0.01)

}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{SAD}
\keyword{abundance,}
\keyword{logseries,}
\keyword{species}
