% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sadMethods.R
\name{plot.sad}
\alias{plot.sad}
\title{Plot method for objects of class \code{sad}}
\usage{
\method{plot}{sad}(x, ptype = c("cdf", "rad"), thr.col = "red", ...)
}
\arguments{
\item{x}{an object of class \code{sad}}
}
\value{
\code{NULL}.
}
\description{
Plot objects of class \code{sad}
}
\details{
See examples
}
\examples{
x <- rtpois(1000, 2)
sad1 <- fitSAD(x, models='tpois')
plot(sad1[[1]])

}
\seealso{
fitSAD
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{SAD}
\keyword{abundance,}
\keyword{species}
