% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redim_mtx.R
\name{redim_mtx}
\alias{redim_mtx}
\title{Create 2-dimensional embedding of a matrix}
\usage{
redim_mtx(mtx, ...)
}
\arguments{
\item{mtx}{The matrix to be embedded in 2 dimension.}

\item{...}{Any additional arguments.}
}
\value{
A list including the 2-dimensional embedded dataset and the plot if \code{plot==TRUE}.
}
\description{
Creates 2-dimensional embedding of a matrix and visualize the result.
}
\section{Additional arguments}{

\describe{
\item{method}{A string indicating the dimensionality reduction method (possible values = \code{c('pca', 'tsne', 'umap')}).}
\item{n_neighbors}{Only if \code{method == 'umap'}, \code{n_neighbors} passed to \code{uwot::umap}.}
\item{perplexity}{Only if \code{method == 'tsne'}, \code{perplexity} passed to \code{Rtsne::Rtsne}.}
\item{colors}{A vector of the same length as the number of columns, indicating point colors in the 2-dimensional visualization.}
\item{shapes}{A vector of the same length as the number of columns, indicating  point shapes in the 2-dimensional visualization.}
\item{plot}{A logical indicating if the 2-dimensional embedding should be visualized.}
\item{plot_args}{Visualization arguments to be passed to \code{plot_redim_mtx}.}
\item{log}{A logical indicating whether to log the computation times.}
\item{verbose}{A logical indicating whether to print out the computation steps.}
}
}

\author{
Elyas Heidari
}
