% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redim_sce.R
\name{redim_sce}
\alias{redim_sce}
\title{A wrapper for \code{scater}'s dimension reduction functions.}
\usage{
redim_sce(
  sce,
  method = "umap",
  label = c("cell_type", "batch"),
  plot = FALSE,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment to perform dimension reduction on.}

\item{method}{A string indicating the method to be used for dimension reduction.}

\item{label}{Which labels to be used (should be column name from the \code{colData}) for annotating the cells in the visualization.}

\item{plot}{A logical, whether to plot the low-dimensional embedding.}

\item{...}{Any additional arguments.}
}
\value{
A SingleCellExperiment object with the reduced dimention matrix in \code{reducedDims(..., name)}.
}
\description{
Implements PCA, t-SNE and UMAP, as a wrapper for \code{scater}'s dimension reduction functions.
}
\author{
Elyas Heidari
}
