% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.sad}
\alias{residuals.sad}
\alias{mse}
\alias{mse.sad}
\alias{mseZ}
\alias{mseZ.sad}
\title{Compute residuals and mean squared error between observed and predicted SADs.}
\usage{
\method{residuals}{sad}(x, type = c("rank", "cumulative"), relative = TRUE,
  log = FALSE, ...)

mse(x, ...)

\method{mse}{sad}(x, ...)

mseZ(x, ...)

\method{mseZ}{sad}(x, nrep, return.sim = FALSE, ...)
}
\arguments{
\item{x}{an \code{sad} object}

\item{type}{'rank' or 'cumulative'}

\item{relative}{logical, if true use relative MSE; defaults to TRUE}

\item{log}{logical, if TRUE calculate MSE on log scale; defaults to FALSE}

\item{...}{arguments to be passed to methods}

\item{nrep}{number of simulations from the fitted METE distribution}

\item{return.sim}{logical; return the simulated liklihood values}
}
\value{
\describe{
 \item{\code{residuals}}{returns a numeic vector giving residuals for each data point; 
 for \code{type='rank'} this will be a vector of length equal to the number of species 
 but for \code{type='cumulative'} this will be a vector of length equal to the number 
 of unique abundances.}
 \item{\code{mse}}{returns a numeric vector of length one giving the value of the mean 
 squared error.}
 \item{\code{mseZ}}{returns a list with elements: \code{z}, the z-score; \code{obs}, 
 the observed mean squared error; \code{sim}, \code{nrep} simulated values (\code{NULL} 
 if \code{return.sim=FALSE}.)}
}
}
\description{
\code{residuals.sad} computes residuals between predicted and observed SAD, either the rank abundance or CDF. 
\code{mse.sad} computes mean squared error for rank or cdf between predicted and observed SAD.
\code{mseZ.sad} computes the z-score of mean squared error based on parametric bootstrapping.
}
\details{
Residuals are calculated either for the rank abundances or the cumulative distribution function. 
This can be done on a log scale and/or as the proportional error.  The mean squared error is then 
simply the mean of squared residuals of either the rank or cumulative distribution function. 
z-scores are caculated by simulating from a fitted SAD and calculating the MSE between the simulated 
data sets and the SAD prediction. The distribution of these values is compared against the MSE of the 
data to obtain a z-score.
}
\examples{
'stub'

}
\author{
Andy Rominger <ajrominger@gmail.com>
}
