% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sad2Rank.R
\name{sad2Rank}
\alias{sad2Rank}
\title{meteDist2Rank}
\usage{
sad2Rank(x, S = NULL)
}
\arguments{
\item{x}{\code{meteDist} object}
}
\value{
A vector of predicted ranked abundances
}
\description{
Calculate the rank distribution of a sad object
}
\details{
Extracts the predicted rank distribution from a \code{sad} object.
This is effectively the quantile function of the distribution.
}
\examples{
'stub'
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\seealso{
fitSAD, qqplot
}

