% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSAD.R
\name{sample.sad}
\alias{sample.sad}
\title{Sample from an SAD object}
\usage{
sample.sad(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{an object of class \code{sad}}

\item{size}{the number of individuals to sample from the SAD}

\item{replace}{logical, should sampling be done with replacement or without}

\item{prob}{an optional numeric specifying probability of inclusion in a sample, either a 
single scalar value or vector of length equal to length of \code{x}, see Details.}
}
\value{
A numeric vector of variable length depending on how many species have sampled abundance > 0
}
\description{
Samples a desired number of individuals from an \code{sad} object.
}
\details{
If a single scalar values is supplied for \code{prob}, size will be 
calculated as \code{size = prob*sum(x)}; if a vector of length equal to length of \code{x} is giving this will
not change the value of \code{size} but rather indicate the relative probabilities of each species
appearing in the sample (e.g. for simulating species-specific detection bias)
}
\examples{
x <- rtpois(1000, 2)
sad1 <- fitSAD(x, models='tpois')
sample(sad1[[1]], 10)

}
\seealso{
sample
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{SAD}
\keyword{abundance,}
\keyword{species}
\keyword{subsampling,}
