% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brokenStick.R
\name{dstick}
\alias{dstick}
\alias{pstick}
\alias{qstick}
\alias{rstick}
\title{Broken stick}
\usage{
dstick(x, r, log = FALSE)

pstick(q, r, lower.tail = TRUE, log = FALSE)

qstick(p, r, lower.tail = TRUE, log = FALSE)

rstick(n, r)
}
\arguments{
\item{x}{vector of integers for which to return the probability}

\item{r}{parameter of the broken stick model, see Details}

\item{log}{logical, should the log probability be used}

\item{q}{vector of integers for which to return the cumulative probability}

\item{lower.tail}{logical, should the lower tail be used}

\item{p}{vector of probabilities for which to return the quantile}

\item{n}{number of random replicates}
}
\value{
A numeric vector of length equal to the input
}
\description{
\code{dstick} gives the probability mass function, \code{pstick} gives the cumulative mass
function, \code{qstick} the quantile function, \code{rstick} randome number generation
}
\details{
These functions assume infinite support of the Broken stick model. Parameter \code{r} corresponds roughly to the probability of dividing the resource, but really should be interpreted in terms of the underlying PMF: \eqn{(1-r)^{x-1}r}
}
\examples{
dstick(1:10, 0.5)
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\seealso{
fitSAD, dgeom
}
\keyword{Broken}
\keyword{SAD}
\keyword{abundance,}
\keyword{distribution,}
\keyword{geometric}
\keyword{species}
\keyword{stick,}

