% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_cells.R
\name{subset_cells}
\alias{subset_cells}
\title{Cell subset selection}
\usage{
subset_cells(sce, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment to select a subset of cells from.}

\item{...}{Any additional arguments.}
}
\value{
The SingleCellExperiment object with the subset of cells selected.
}
\description{
Subsets cells from the SingleCellExperiment object.
}
\section{Additional arguments}{

\describe{
\item{cell_types}{a string vector including a subset of cell types. Set \code{cell_types='all'} to select from all cell types.}
\item{min_num}{minimum number of cells (in the train dataset) for a cell type to be selected.}
\item{min_rel_prop}{minimum reletive proportion of number of cells (in the train dataset) for a cell type to be selected.}
\item{balanced}{a logical indicating if subsampling should be balanced (i.e., each cell type).}
}
}

