% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genes.R
\name{subset_genes}
\alias{subset_genes}
\title{Gene subset selection}
\usage{
subset_genes(sce, ...)
}
\arguments{
\item{sce}{A SingleCellExperiment to select a subset of genes from.}

\item{...}{Any additional arguments.}
}
\value{
The SingleCellExperiment object with the subset of genes selected.
}
\description{
Subsets genes to be used as the network nodes
}
\section{Additional arguments}{

\describe{
\item{n}{An integer indicating the number of genes to be selected.}
\item{method}{selection based on transcription factors, highly variable genes, or randomly or a combination of these. (possible values = \code{c('tf', 'hvg', 'rand')}).}
}
}

