% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_net.R
\name{tf_net}
\alias{tf_net}
\title{Construct the transcription factor network}
\usage{
tf_net(...)
}
\arguments{
\item{...}{Any additional arguments.}
}
\value{
The adjacency matrix as a sparse matrix of type \code{dgCMatrix}.
}
\description{
Constructs the transcription factor interaction network based on RegNetwork.
}
\section{Additional arguments}{

\describe{
\item{spec}{A character string indicating the organism. possible values = \code{c('cattle', 'chicken','chimpanzee', 'dog_domestic', 'frog_western clawed', 'human', 'mouse_laboratory','rat', 'zebrafish')}.}
\item{tf_subset}{A vector of character strings indicating the transcription factors to be selected.}
\item{log}{A logical indicating whether to log the computation times.}
\item{verbose}{A logical indicating whether to print out the computation steps.}
}
}

\references{
Zhi-Ping Liu, Canglin Wu, Hongyu Miao and Hulin Wu (2015). RegNetwork: an integrated database of transcriptional and posttranscriptional regulatory networks in human and mouse. Database 2015. doi: 10.1093/database/bav095

@author Elyas Heidari
}
