% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/truncNegBinom.R
\name{dtnegb}
\alias{dtnegb}
\alias{ptnegb}
\alias{qtnegb}
\alias{rtnegb}
\title{Truncated negative binomial distribuiton}
\usage{
dtnegb(x, mu, k, log = FALSE)

ptnegb(q, mu, k, lower.tail = TRUE, log = FALSE)

qtnegb(p, mu, k, lower.tail = TRUE, log = FALSE)

rtnegb(n, mu, k)
}
\arguments{
\item{x}{vector of integers for which to return the probability}

\item{mu}{mean log abundance, see Details}

\item{log}{logical, should the log probability be used}

\item{q}{vector of integers for which to return the cumulative probability}

\item{lower.tail}{logical, should the lower tail be used}

\item{p}{vector of probabilities for which to return the quantile}

\item{n}{number of random replicates}

\item{sig}{standard deviation of log abundance, see Details}
}
\value{
A numeric vector of length equal to the input
}
\description{
\code{dptnegb} gives the probability mass function, \code{ptnegb} gives the cumulative mass
function, \code{qtnegb} the quantile function, \code{rtnegb} randome number generation
}
\details{
These functions assume infinite support of the Poisson log-normal from [1, Inf).
}
\examples{
dtnegb(1:10, 0.5, 2)
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{SAD}
\keyword{Truncated}
\keyword{abundance,}
\keyword{binomial,}
\keyword{negative}
\keyword{species}

