% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncPois.R
\name{dtpois}
\alias{dtpois}
\alias{ptpois}
\alias{qtpois}
\alias{rtpois}
\title{Truncated Poisson distribuiton}
\usage{
dtpois(x, lambda, log = FALSE)

ptpois(q, lambda, lower.tail = TRUE, log = FALSE)

qtpois(p, lambda, lower.tail = TRUE, log = FALSE)

rtpois(n, lambda)
}
\arguments{
\item{x}{vector of integers for which to return the probability}

\item{lambda}{mean abundance}

\item{log}{logical, should the log probability be used}

\item{q}{vector of integers for which to return the cumulative probability}

\item{lower.tail}{logical, should the lower tail be used}

\item{p}{vector of probabilities for which to return the quantile}

\item{n}{number of random replicates}
}
\value{
A numeric vector of length equal to the input
}
\description{
\code{dtpois} gives the probability mass function, \code{ptpois} gives the cumulative mass 
function, \code{qtpois} the quantile function, \code{rtpois} randome number generation
}
\details{
These functions assume infinite support of the truncated Poisson from [1, Inf).
}
\examples{

dtpois(1:10, 2)

}
\seealso{
dpois
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{Poisson,}
\keyword{SAD}
\keyword{Truncated}
\keyword{abundance,}
\keyword{species}
