% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{calc_percent_change}
\alias{calc_percent_change}
\title{This function converts a count variable over time into a percent change based on the average
value in the specified baseline period. This function was written for application to
mobility data, where the percent change in mobility over time relative to baseline is of
interest. However, this function can be applied to any count type time series.}
\usage{
calc_percent_change(
  dat,
  date_var = "date",
  grp_var,
  count_var,
  n_baseline_periods = 7,
  start_date = NULL
)
}
\arguments{
\item{dat}{data frame with columns that correspond to count variable and grouping variable}

\item{date_var}{character string of the name of the date column (should be of class "Date")}

\item{grp_var}{character string of column name in dat to be used as grouping variable}

\item{count_var}{character string of the name of the count column, such as number of trips}

\item{n_baseline_periods}{Number of periods to calculate baseline average over. For example,
if the time series is days, n_baseline_periods = 7 for a baseline week.}

\item{start_date}{start date of baseline period (character string)}
}
\value{
data frame of with an additional column of the percent change relative to baseline
}
\description{
This function converts a count variable over time into a percent change based on the average
value in the specified baseline period. This function was written for application to
mobility data, where the percent change in mobility over time relative to baseline is of
interest. However, this function can be applied to any count type time series.
}
\keyword{pika}
