% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{cross_corr}
\alias{cross_corr}
\title{This function determines lag at which the cross correlation is highest}
\usage{
cross_corr(dat, date_var = NULL, grp_var, x_var, y_var, max_lag = 20,
  subset_date = NULL)
}
\arguments{
\item{dat}{data frame with columns that correspond to two time series and grouping variable(s)}

\item{date_var}{character string of the column name that corresponds to the date variable
date_var must be specified if subset_date is not null.}

\item{grp_var}{character string of column names in dat to be used as grouping variable(s)}

\item{x_var}{primary time series (should be a column in dat)}

\item{y_var}{secondary time series (should be a column in dat)}

\item{max_lag}{integer value of the number of lags to perform cross correlation for}

\item{subset_date}{a character string of the same format as the date variable}
}
\value{
tibble of lags by grp_var
}
\description{
This function determines lag at which the cross correlation is highest
}
\keyword{pika}
