% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{estimate_rt}
\alias{estimate_rt}
\title{Estimate reproduction over time by group}
\usage{
estimate_rt(
  dat,
  grp_var,
  date_var,
  incidence_var,
  est_method = "parametric_si",
  si_mean = 6.48,
  si_std = 3.83
)
}
\arguments{
\item{dat}{data frame with columns that correspond to two time series and grouping variable(s)}

\item{grp_var}{character string of column names in dat to be used as grouping variable(s)}

\item{date_var}{primary time series (should be a column in dat)}

\item{incidence_var}{secondary time series (should be a column in dat)}

\item{est_method}{estimation method to be used to estimate R in EpiEstim:
c("non_parametric_si","parametric_si","uncertain_si","si_from_data","si_from_sample")}

\item{si_mean}{mean of serial interval distribution to be specified when
est_method = "parametric_si"}

\item{si_std}{standard deviation of serial interval distribution to be specified when
est_method = "parametric_si"}
}
\value{
data frame of mean, median, and 95% credible interval reproduction number
}
\description{
This function estimates reproduction number by group using EpiEstim's estimate_R() and
then binds the results together into a single data frame
}
\keyword{pika}
