% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_lag}
\alias{plot_lag}
\title{This function plots the lags by grp_var}
\usage{
plot_lag(dat, lag_var, bins = 2)
}
\arguments{
\item{dat}{data frame with columns that correspond to two time series and grouping variable(s)}

\item{lag_var}{character string of the column name that corresponds to the date variable}

\item{bins}{character string of column names in dat to be used as grouping variable(s)}
}
\value{
tibble of lags by grp_var
}
\description{
This function plots the lags by grp_var
}
\keyword{pika}
