% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/poissonLogNorm.R
\name{dplnorm}
\alias{dplnorm}
\alias{pplnorm}
\alias{qplnorm}
\alias{rplnorm}
\title{Poisson log-normal}
\usage{
dplnorm(x, mu, sig, log = FALSE)

pplnorm(q, mu, sig, lower.tail = TRUE, log = FALSE)

qplnorm(p, mu, sig, lower.tail = TRUE, log = FALSE)

rplnorm(n, mu, sig)
}
\arguments{
\item{x}{vector of integers for which to return the probability}

\item{mu}{mean log abundance, see Details}

\item{sig}{standard deviation of log abundance, see Details}

\item{log}{logical, should the log probability be used}

\item{q}{vector of integers for which to return the cumulative probability}

\item{lower.tail}{logical, should the lower tail be used}

\item{p}{vector of probabilities for which to return the quantile}

\item{n}{number of random replicates}
}
\value{
A numeric vector of length equal to the input
}
\description{
\code{dplnorm} gives the probability mass function, \code{pplnorm} gives the cumulative mass
function, \code{qplnorm} the quantile function, \code{rplnorm} randome number generation
}
\details{
These functions assume infinite support of the Poisson log-normal from [1, Inf).
}
\examples{
dplnorm(1:10, 0.5, 0.1)
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{Poisson}
\keyword{SAD}
\keyword{abundance,}
\keyword{log}
\keyword{normal,}
\keyword{species}

