% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sadMethods.R
\name{print.sad}
\alias{print.sad}
\title{Print method for objects of class \code{sad}}
\usage{
\method{print}{sad}(x)
}
\arguments{
\item{x}{an object of class \code{sad}}
}
\value{
Silently returns \code{x}.
}
\description{
Print objects of class \code{sad}
}
\details{
See examples
}
\examples{
x <- rtpois(1000, 2)
sad1 <- fitSAD(x, models='tpois')
print(sad1[[1]])

}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\seealso{
fitSAD
}
\keyword{SAD}
\keyword{abundance,}
\keyword{species}

