% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{rolling_corr}
\alias{rolling_corr}
\title{Calculate rolling correlation between two time series by group}
\usage{
rolling_corr(dat, date_var = "date", grp_var, x_var, y_var, n = 14)
}
\arguments{
\item{dat}{data frame with columns that correspond to two time series and grouping variable}

\item{date_var}{character string of date column name (should be of class "Date")}

\item{grp_var}{character string of column name in dat to be used as grouping variable}

\item{x_var}{primary time series (should be a column in dat)}

\item{y_var}{secondary time series (should be a column in dat)}

\item{n}{the number of time periods over which to calculate rolling correlation}
}
\value{
tibble of lags by grp_var
}
\description{
This function calculates the rolling correlation between two time series
}
\keyword{pika}
