% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{rolling_corr}
\alias{rolling_corr}
\title{This function calculates the rolling correlation between two time series}
\usage{
rolling_corr(dat, grp_var, x_var, y_var, n = 14)
}
\arguments{
\item{dat}{data frame with columns that correspond to two time series and grouping variable(s)}

\item{grp_var}{character string of column names in dat to be used as grouping variable(s)}

\item{x_var}{primary time series (should be a column in dat)}

\item{y_var}{secondary time series (should be a column in dat)}

\item{n}{the number of time periods over which to calculate rolling correlation}
}
\value{
tibble of lags by grp_var
}
\description{
This function calculates the rolling correlation between two time series
}
\keyword{pika}
